Move shared inner classes to separate files
